Updated to a less verbose (horse-shit) dictionary
Also removed many dictionary files that were slowing down the grunt serve.

Also fixed a critical error in filtering the word list, where in if a word was not in the word list already, it could not be added to that list. That was a problem. I fixed it.

Just a side note: the new dictionary came from "http://www.mieliestronk.com/wordlist.html"
It's british, 58K words, and is tailored to crossword puzzle words. So certain phrases are also included.
The trade-off is that most of the words it finds are words a normal person would recognize.